+++ packages/subgraph/src/registry.ts
@@ -1,8 +1,6 @@
 import {NewDAORegistered} from '../generated/Registry/Registry';
-import {DAO as DAOContract} from '../generated/templates/DAO/DAO';
-import {DAO} from '../generated/templates';
+import {DaoTemplate} from '../generated/templates';
 import {Dao} from '../generated/schema';
-import {DataSourceContext} from '@graphprotocol/graph-ts';
 
 export function handleNewDAORegistered(event: NewDAORegistered): void {
   let id = event.params.dao.toHexString(); // use dao address as id, because it should not repeat
@@ -13,7 +11,7 @@ export function handleNewDAORegistered(event: NewDAORegistered): void {
   entity.token = event.params.token;
 
   // subscribe to templates
-  DAO.create(event.params.dao);
+  DaoTemplate.create(event.params.dao);
 
   entity.save();
 }
